# api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package api

import (
	"time"

	"github.com/labstack/echo/v4"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "BearerAuth.Scopes"
)

// AuthenticationRequest defines model for AuthenticationRequest.
type AuthenticationRequest struct {
	// Name ユーザーの名前
	Name string `json:"name"`

	// Password パスワード
	Password string `json:"password"`
}

// AuthenticationResponse defines model for AuthenticationResponse.
type AuthenticationResponse struct {
	// Name ユーザーの名前
	Name *string `json:"name,omitempty"`

	// Token JWTトークン
	Token *string `json:"token,omitempty"`

	// Uid 10000から始まるUID
	Uid *int `json:"uid,omitempty"`
}

// DeleteTaskRequest defines model for DeleteTaskRequest.
type DeleteTaskRequest struct {
	// Id タスクのID
	Id int `json:"id"`
}

// NewTaskRequest defines model for NewTaskRequest.
type NewTaskRequest struct {
	// Assign 担当者のuid
	Assign *openapi_types.UUID `json:"assign,omitempty"`

	// Deadline 真の締切（ISO8601形式を推奨）
	Deadline time.Time `json:"deadline"`

	// Group タスクグループのID
	Group *openapi_types.UUID `json:"group,omitempty"`

	// Name 名前
	Name string `json:"name"`

	// Period 目安の締切
	Period *time.Time `json:"period,omitempty"`

	// Priority 優先度
	Priority int `json:"priority"`

	// Tag ジャンル
	Tag string `json:"tag"`
}

// TaskResponse defines model for TaskResponse.
type TaskResponse struct {
	// Assign 担当者のuid
	Assign *openapi_types.UUID `json:"assign,omitempty"`

	// Deadline 真の締切（ISO8601形式を推奨）
	Deadline time.Time `json:"deadline"`

	// Group タスクグループのID
	Group *openapi_types.UUID `json:"group,omitempty"`

	// Id タスクのID
	Id *int `json:"id,omitempty"`

	// Name 名前
	Name string `json:"name"`

	// Period 目安の締切
	Period *time.Time `json:"period,omitempty"`

	// Priority 優先度
	Priority int `json:"priority"`

	// Tag ジャンル
	Tag string `json:"tag"`
}

// UpdateTaskRequest defines model for UpdateTaskRequest.
type UpdateTaskRequest struct {
	// Assign 担当者のuid
	Assign *openapi_types.UUID `json:"assign,omitempty"`

	// Deadline 真の締切（ISO8601形式を推奨）
	Deadline *time.Time `json:"deadline,omitempty"`

	// Group タスクグループのID
	Group *openapi_types.UUID `json:"group,omitempty"`

	// Id タスクのID
	Id int `json:"id"`

	// Name 名前
	Name *string `json:"name,omitempty"`

	// Period 目安の締切
	Period *time.Time `json:"period,omitempty"`

	// Priority 優先度
	Priority *int `json:"priority,omitempty"`

	// Tag ジャンル
	Tag *string `json:"tag,omitempty"`
}

// Accepted defines model for Accepted.
type Accepted struct {
	Message *string `json:"message,omitempty"`
}

// BadRequest defines model for BadRequest.
type BadRequest struct {
	Message *string `json:"message,omitempty"`
}

// InternalServerError defines model for InternalServerError.
type InternalServerError struct {
	Message *string `json:"message,omitempty"`
}

// DeleteTaskJSONRequestBody defines body for DeleteTask for application/json ContentType.
type DeleteTaskJSONRequestBody = DeleteTaskRequest

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = AuthenticationRequest

// CreateTaskJSONRequestBody defines body for CreateTask for application/json ContentType.
type CreateTaskJSONRequestBody = NewTaskRequest

// SignupJSONRequestBody defines body for Signup for application/json ContentType.
type SignupJSONRequestBody = AuthenticationRequest

// UpdateTaskJSONRequestBody defines body for UpdateTask for application/json ContentType.
type UpdateTaskJSONRequestBody = UpdateTaskRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// delete task
	// (POST /api/deletetask)
	DeleteTask(ctx echo.Context) error
	// login
	// (POST /api/login)
	Login(ctx echo.Context) error
	// create new task
	// (POST /api/newtask)
	CreateTask(ctx echo.Context) error
	// signup
	// (POST /api/signup)
	Signup(ctx echo.Context) error
	// get all tasks
	// (GET /api/tasks)
	GetTasks(ctx echo.Context) error
	// update task
	// (POST /api/updatetask)
	UpdateTask(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// DeleteTask converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteTask(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteTask(ctx)
	return err
}

// Login converts echo context to params.
func (w *ServerInterfaceWrapper) Login(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.Login(ctx)
	return err
}

// CreateTask converts echo context to params.
func (w *ServerInterfaceWrapper) CreateTask(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateTask(ctx)
	return err
}

// Signup converts echo context to params.
func (w *ServerInterfaceWrapper) Signup(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.Signup(ctx)
	return err
}

// GetTasks converts echo context to params.
func (w *ServerInterfaceWrapper) GetTasks(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetTasks(ctx)
	return err
}

// UpdateTask converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateTask(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UpdateTask(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/api/deletetask", wrapper.DeleteTask)
	router.POST(baseURL+"/api/login", wrapper.Login)
	router.POST(baseURL+"/api/newtask", wrapper.CreateTask)
	router.POST(baseURL+"/api/signup", wrapper.Signup)
	router.GET(baseURL+"/api/tasks", wrapper.GetTasks)
	router.POST(baseURL+"/api/updatetask", wrapper.UpdateTask)

}