# openapi.yaml
openapi: 3.0.3

info:
  title: Uprav API
  version: 1.0.0
  description: Uprav API

servers:
  - url: http://localhost:1124
    description: development server

paths:
  /api/tasks:
    get:
      summary: get all tasks
      tags:
        - Tasks
      operationId: getTasks
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/newtask:
    post:
      summary: create new task
      tags:
        - Tasks
      operationId: createTask
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewTaskRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/updatetask:
    post:
      summary: update task
      tags:
        - Tasks
      operationId: updateTask
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/deletetask:
    post:
      summary: delete task
      tags:
        - Tasks
      operationId: deleteTask
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteTaskRequest"
      responses:
        "200":
          $ref: "#/components/responses/Accepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/login:
    post:
      summary: login
      tags:
        - Authentication
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthenticationRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/signup:
    post:
      summary: signup
      tags:
        - Authentication
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthenticationRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  responses:
    Accepted:
      description: Accepted
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    NoContent:
      description: No Content
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Internal server error
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    NewTaskRequest:
      type: object
      properties:
        name:
          type: string
          description: "名前"
          example: "タスクのタイトル"
        priority:
          type: integer
          description: "優先度"
          minimum: 1
          maximum: 5
          example: 3
        tag:
          type: string
          description: "ジャンル"
          example: "課題"
        deadline:
          type: string
          format: date-time
          description: "真の締切（ISO8601形式を推奨）"
          example: "2023-12-31T23:59:59Z"
        period:
          type: string
          format: date-time
          description: "目安の締切"
          example: "2023-12-25T18:00:00Z"
        # assumption:
        #   type: array
        #   items:
        #     type: string
        #     format: uuid
        #   description: "前提タスクのID(複数を指定可能)"
        group:
          type: string
          format: uuid
          description: "タスクグループのID"
        assign:
          type: string
          format: uuid
          description: "担当者のuid"

      #必須項目がある場合はここに追加します
      required:
        - name
        - priority
        - tag
        - deadline
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          description: "タスクのID"
          example: 1
        name:
          type: string
          description: "名前"
          example: "タスクのタイトル"
        priority:
          type: integer
          description: "優先度"
          minimum: 1
          maximum: 5
          example: 3
        tag:
          type: string
          description: "ジャンル"
          example: "課題"
        deadline:
          type: string
          format: date-time
          description: "真の締切（ISO8601形式を推奨）"
          example: "2023-12-31T23:59:59Z"
        period:
          type: string
          format: date-time
          description: "目安の締切"
          example: "2023-12-25T18:00:00Z"
        # assumption:
        #   type: array
        #   items:
        #     type: string
        #     format: uuid
        #   description: "前提タスクのID(複数を指定可能)"
        group:
          type: string
          format: uuid
          description: "タスクグループのID"
        assign:
          type: string
          format: uuid
          description: "担当者のuid"
      required:
        - name
        - priority
        - tag
        - deadline
    UpdateTaskRequest:
      type: object
      properties:
        id:
          type: integer
          description: "タスクのID"
          example: 1
        name:
          type: string
          description: "名前"
          example: "タスクのタイトル"
        priority:
          type: integer
          description: "優先度"
          minimum: 1
          maximum: 5
          example: 3
        tag:
          type: string
          description: "ジャンル"
          example: "課題"
        deadline:
          type: string
          format: date-time
          description: "真の締切（ISO8601形式を推奨）"
          example: "2023-12-31T23:59:59Z"
        period:
          type: string
          format: date-time
          description: "目安の締切"
          example: "2023-12-25T18:00:00Z"
        # assumption:
        #   type: array
        #   items:
        #     type: string
        #     format: uuid
        #   description: "前提タスクのID(複数を指定可能)"
        group:
          type: string
          format: uuid
          description: "タスクグループのID"
        assign:
          type: string
          format: uuid
          description: "担当者のuid"
      required:
        - id
    DeleteTaskRequest:
      type: object
      properties:
        id:
          type: integer
          description: "タスクのID"
          example: 1
      required:
        - id
    AuthenticationRequest:
      type: object
      properties:
        name:
          type: string
          description: "ユーザーの名前"
          example: "Mira"
        password:
          type: string
          description: "パスワード"
          example: "password123"
      required:
        - name
        - password
    AuthenticationResponse:
      type: object
      properties:
        uid:
          type: integer
          description: "10000から始まるUID"
          example: 10001
        name:
          type: string
          description: "ユーザーの名前"
          example: "Mira"
        token:
          type: string
          description: "JWTトークン"
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."